Rotating and scaling through matrices

+++ src/application/Time.cpp
@@ -4,7 +4,7 @@ Time::Time()
 {
 }
 
-double Time::deltaTime()
+float Time::deltaTime()
 {
 	//auto currentStep = std::chrono::high_resolution_clock::now();
 	//return std::chrono::duration_cast<std::chrono::seconds>(currentStep - prevStep).count();
@@ -18,5 +18,5 @@ void Time::StartFrame()
 
 void Time::EndFrame()
 {
-	prevStep = std::chrono::duration<double>(std::chrono::high_resolution_clock::now() - frameTime).count();
+	prevStep = std::chrono::duration<float>(std::chrono::high_resolution_clock::now() - frameTime).count();
 }
